+++ kroxylicious-operator/src/main/java/io/kroxylicious/kubernetes/operator/StatusFactory.java
@@ -72,13 +72,4 @@ public abstract class StatusFactory<R extends CustomResource<?, ?>> {
     public abstract R newTrueConditionStatusPatch(R observedProxy,
                                                   Condition.Type type,
                                                   String checksum);
-
-    /**
-     * Deprecated in favour of {@link StatusFactory#newTrueConditionStatusPatch(CustomResource, Condition.Type, String)}
-     *
-     * @deprecated
-     */
-    @Deprecated(forRemoval = true)
-    public abstract R newTrueConditionStatusPatch(R observedProxy,
-                                                  Condition.Type type);
 }